+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Plan B 0.2.0
+
+* Added `moveByRenaming` option to allow move files and directories by
+  renaming. This is not default because moving by renaming is not always
+  possible, while moving by coping is always an option.
+
 ## Plan B 0.1.1
 
 * Fixed the problem with moving of files and directories from `/tmp/` to